Revert "Revert "Add Roles To ProjectsController""
This reverts commit d9dfc8b5be9684b96458def5faf74521e312f186.

+++ YouToDo/Controllers/ToDoProjectsController.cs
@@ -21,6 +21,8 @@ namespace YouToDo.Controllers
         [Route("api/projects")]
         public IQueryable<ToDoProject> GetToDoProjects()
         {
+            if (!User.IsInRole("Admin"))
+                return null;
             return db.ToDoProjects;
         }
 
@@ -34,7 +36,8 @@ namespace YouToDo.Controllers
             {
                 return NotFound();
             }
-
+            if (!User.Identity.Name.Equals(toDoProject.ProjectLeader))
+                return BadRequest("You are not project leader of this project");
             return Ok(toDoProject);
         }
 
@@ -52,6 +55,9 @@ namespace YouToDo.Controllers
             {
                 return BadRequest();
             }
+
+            if (!User.Identity.Name.Equals(toDoProject.ProjectLeader))
+                return BadRequest("You are not project leader of this project");
 
             db.Entry(toDoProject).State = EntityState.Modified;
 
@@ -84,6 +90,9 @@ namespace YouToDo.Controllers
                 return BadRequest(ModelState);
             }
 
+            if ((!User.IsInRole("Manager")) && (!User.IsInRole("Admin")))
+                return BadRequest("You are not a Manager");
+
             db.ToDoProjects.Add(toDoProject);
             db.SaveChanges();
 
@@ -100,6 +109,9 @@ namespace YouToDo.Controllers
             {
                 return NotFound();
             }
+
+            if ((!User.Identity.Name.Equals(toDoProject.ProjectLeader)) && (!User.IsInRole("Admin")))
+                return BadRequest("You are not project leader of this project");
 
             db.ToDoProjects.Remove(toDoProject);
             db.SaveChanges();
